commit callsteck devel

+++ src/callstack.ts
@@ -74,23 +74,29 @@ const bootstrap = async (config: Config) => {
       const method = `get${entity.name}` as keyof Repository;
       let page = 1;
       let response = await respository[method]({ limit, page }, "T");
+      if (response) {
+        //console.log(response);
+        Logger.info("CRIANDO ARQUIVO PARA TRANSMISSAO");
 
-      Logger.info("CRIANDO ARQUIVO PARA TRANSMISSAO");
+        while (0 < response.length) {
+          await csv.write(
+            file,
+            response.map((row: DatabaseRow) => Hydrator(dto, row))
+          );
+          page++;
+          response = await respository[method]({ limit, page }, "T");
+        }
 
-      while (0 < response.length) {
-        await csv.write(
-          file,
-          response.map((row: DatabaseRow) => Hydrator(dto, row))
+        Logger.info("ENVIANDO DADOS VIA SFTP");
+        await ftp.sendFile(entity.file, file);
+
+        if (fs.existsSync(file)) {
+          fs.unlinkSync(file);
+        }
+      } else {
+        Logger.warning(
+          `ERRO AO BUSCAR DADOS NO REPOSITORIO ${entity.name.toLocaleUpperCase()}`
         );
-        page++;
-        response = await respository[method]({ limit, page }, "T");
-      }
-
-      Logger.info("ENVIANDO DADOS VIA SFTP");
-      await ftp.sendFile(entity.file, file);
-
-      if (fs.existsSync(file)) {
-        fs.unlinkSync(file);
       }
     }
